Allow Cancelling Multiple Orders At Once (#738)
This PR adds multiple order cancellation to the orderbook.

The context for this change is that the CoW Swap limit order front end wants to add support for cancelling multiple orders at once.

With the current API (`DELETE /api/v1/orders/${orderUid}`), this has to be done where, for each order, a wallet signature is requested. This means that if you were to cancel 10 orders, you would need to sign 10 MetaMask popups at once. In order to work around this, we introduce a new `OrderCancellations` message that can be signed to indicate intent to cancel multiple orders at once. In the same example as above, the user would only need to sign a single MetaMask popup (including the order UIDs of all 10 orders).

The actual API documentation and end point will be added in a follow up PR, this just contains the implementation in the inner `Orderbook` component.

### Test Plan

CI, added a couple unit tests to smoke test the logic. Want to include an E2E test when the API endpoint gets introduced in the follow-up PR.