Refactor Settlement Encoding to Use Single Trade Type (#779)
Fixes #777.

This PR fixes #777 by refactoring settlement encoding to use a single `Trade` type. This is done by introducing a new `EncoderTrade` type that has a `TokenReference` enum (and is used for differentiating between market trades where both tokens are indexed in the UCP vector, and custom price trades where only the sell token is).

By using a single `Vec` of trades, it makes accidentally considering limit orders as user orders in some cases but not others less likely, and fixes the issue where we were not adding limit order surplus to the total surplus.

Note that there is some additional cleanup that can be done (such as unifying surplus computation functions - I think we have 3 different ones right now), but that will be for a follow up PR.

The PR ended up getting quite large. More than half of the changed lines are adjusting unit tests to build and pass. The first commit in the PR has most of the logic changes.

### Test Plan

Adjusted existing unit tests to build and pass. Note that we more actively enforce invariants when constructing test `Settlement` instances, so it mostly involved adding the correct `executed_amount` values and missing token prices.