Add Support for Limit Order Buying Eth (#798)
This PR fixes an issue with encoding limit orders with the special `BUY_ETH_ADDRESS`.

Specifically, we don't add a token equivalency for these kinds of orders (because their buy price) which causes a "missing buy token error" when adding the limit order to the settlement encoder.

I changed the PR to unconditionally add token equivalency for all orders. I did this so that the order converter doesn't need to know implementation details about the `SettlementEncoder`, and even if we add an additional price here that isn't needed, it gets cleaned up in the `drop_unnecessary_tokens_and_prices` call so we don't have to pay more gas for it.

### Test Plan

Added case in unit test.