Change Default Market Makable Token List (#823)
Related to #820

We recently changed the batch auction model preparation to fetch balances for all market makable tokens. This isn't an issue for our deployments because the list is quite small, however, the list that the driver uses by default has 4900(!) tokens.

This PR is just a really easy change to help the situation by:
- using a much smaller list with 50 tokens
- using the actual list that is being used for market makable tokens

### Test Plan

Run locally and see that the new default list is accepted by the driver.