Configurable Block Current Block Fetching Strategy (#825)
Follow up to #821 

As @vkgnosis pointed out, using `eth_call` for fetching the current block number and hash is a work-around for Nethermind specifically (see referenced PR ☝️ for additional context).

However, that PR also replaced the `BlockRetrieving::current_block` in a way that simplified code that needed to fetch the current block (i.e. it dealt with things like missing block number and hash internally, instead of at every call-site).

This PR keeps that refactor and expands on it, refactoring the codebase to use an `Arc<dyn BlockRetrieving>` instead of `Web3` everywhere. This allows us to create the block retriever on start-up and parameterize it. Specifically, there is a block retrieving strategy command line argument that controls whether to use `eth_getBlockAt*` or `eth_call` in order to fetch the current block info.

### Test Plan

CI - existing tests continue to pass.

### Release Notes

We should enable `BLOCK_STREAM_RETRIEVER_STRATEGY=eth-call` to reduce the number of "missing block hash" errors we get from our node.