Fixing conflict in reorg handling by autopilot for ethflow (#964)
Currently, we see on staging the following error:

WARN maintenance{block=16189813}: shared::maintenance: Service Maintenance Error: insert_orders failedCaused by:
    0: error returned from database: current transaction is aborted, commands ignored until end of transaction block
    1: current transaction is aborted, commands ignored until end of transaction block
I think this is due to the fact that a first insert_order returns an error - duplication of uid - which we ignore. Then we try to proceed with another query. But since it's a db-transaction, and not a single query, it keeps on telling that the last command produced already an error.
This PR makes sure that the first insertion does not even fail.
(see here for more context: https://stackoverflow.com/questions/10399727/psqlexception-current-transaction-is-aborted-commands-ignored-until-end-of-tra)