Use block_timestamp instead of system time for onchain order creation date (#968)
In the event, that we have an issue with autopilot order indexing, and we would index orders delayed, there could be a bigger divergence between the order creation onchain and the order indexing CPU time. In order to reflect better the real-time of mining a transaction, we introduce this change. This helps then also later the refunder to make fairer judgments on the actual order_validity_duration for its refunding considerations.