CoW Protocol in the Multicall of Madness (#824)
First off, yes, this is a play on words on the latest Dr. Strange movie (I'm a dad, so dad jokes are allowed 😛) and one small part of #452.

This PR adds a new `multicall` module in order to combine multiple `eth_call`s into a single one that executes a "multicall". This is the concept of using a contract that will execute within a single call context multiple calls and return their joint results, as an alternative to batching RPC `eth_call`s together.

Note that these are more limiting (in that not all call parameters are configurable from the EVM, so things like `gas_price` are shared across all inner calls) but for a large portion of our use cases (such as getting balances and pool reserves for 10's and sometimes 100's of tokens) is enough.

The strategy we use for executing multicalls here is non-conventional. Typically, the [`Multicall2`](https://github.com/makerdao/multicall/blob/master/src/Multicall2.sol) contract is used. This contract has been deployed on multiple chains (including all the ones that we care about), but I did not like relying on a deployed contract for just reading blockchain state (it would also require us to deploy it to our local node for E2E tests). Instead, what we do is execute a contract creation call, where the resulting code of the new contract is the return data we care about. `eth_call` allows this will gladly return the data that we encode (tested on Nethermind and Geth, didn't have access to an Erigon node for testing but would be surprised if it didn't work there as well). This allows us to simulate arbitrary EVM bytecode execution without actually needing an on-chain contract to support it :tada:. 

With this PR, we just add some methods for facilitating these kinds of calls, but don't make use of them (that will come separately). The plan is to eventually automatically batch multiple `eth_call`s into a single multicall instead of including them in a regular batch. This gives us:
- State consistency across the multicalls
- Lowers number of `eth_call`s which decreases external node provider expenses

### Test Plan

Added unit tests! Also some manual tests. You can see [a Tenderly simulation](https://dashboard.tenderly.co/gp-v2/staging/simulator/9cf06ce9-9336-4b9b-b4ea-7a3755524c24) of what one of these multicalls actually looks like (this is the multicall from the `execute_multicall` unit test).