Prevent limit order fee updater from getting stuck (#965)
This is a theoretical issue I noticed while thinking about limit orders as part of oncall.

Currently, the limit order fee updater retrieves orders with outdated fees in an arbitrary ordering because the sql query does not include an ORDER BY clause. This is problematic because it is possible that the fee estimate for an order keeps failing forever, for example because the quote is out of price. If this happened for all of the 100 orders that the query returns, then the updater would never make progress.

This PR fixes this issue by slightly changing the semantics of the surplus_fee and surplus_fee_timestamp columns. When updating the fee fails, surplus_fee is set to null and surplus_fee_timestamp is updated to the current time. With this change we can return the limit orders with outdated fees ordered by surplus_fee_timestamp, giving us the most outdated orders first.

As part of this I am categorizing errors into preventable and unpreventable to improve logging and metrics.