Index refund events of ethflow and store the tx hash (#970)
Fixes #803

This PR implements the required changes to move from an is_refunded flag for ethflow orders to storing the transaction hash that refunded the order.

added new components that listen for and index only refund events from the ethflow. This seemed like the right approach because the other ethflow events are already indexed by the generic CoWSwapOnChainOrderContract interface and indexing the refund events as part of that didn't seem possible.

added a new table ethflow_refunds

order_uid as a private key to quickly associate a refund with the corresponding order
tx_hash to have the data we are interested in
bock_number to be able to deal with reorgs
I split EthOrderPlacement into 2 types (one for creating the order, and one for reading the order which contains the refund_tx)

I added refundTxHash to the openapi spec (isRefunded was kept for backwards compatibility reasons but only implemented with custom serialization logic)

added a DB migration script to create placeholder entries in ethflow_refunds for all ethflow orders with is_refunded = true. Those placeholders can later be replaced with the correct values with a manual update.