Fix not filtering out already solved orders for limit order quoting (#990)
The queries dealing with limit orders are not taking into account all the conditions that can cause an order to be unsolvable. Most importantly the "already settled" condition was missing.

In this PR I factor out the solvable orders query so that it can be reused by the solvable_orders function and the limit order related queries. This ensures that both work on the same conditions. Reusing this query is becoming hairy. If these queries diverge more I would stop reusing and copy paste instead.

After merging we should look at the speed of the changed queries to see if there is problematic performance.