Include Uninternalized Settlement In Solver Competition Data (#991)
Fixes https://github.com/cowprotocol/services/issues/604

This PR adds a new `uniniternalized_call_data` field to the settlement
competition debug JSON blob that gets uploaded to the database.

The motivation behind this change is that, currently, the only way for
external solvers to inspect internalized data from their competition is
to ask us to look it up in the logs. This is not ideal.

For this change, I made the `uniniternalized_call_data` **only get added
if it different from `call_data`**. The reason for this is database size
considerations.

### Test Plan

Rust compiler. Serialization test.