Improve panic logging (#992)
When we created the current panic handler, it was not possible to
manually print the stacktrace. We relied on logging an error message
first and the panic being printed immediately after.

With one of the recent rust versions it became possible to emulate the
output of the default panic handler. This has the advantage that we can
form a single proper log message with the panic message.

I noticed that sometimes panics don't correctly show up in Kibana, which
is fixed by this PR.

### Test Plan

manual test of panic output

To test you can run `cargo test --package shared --lib --
exit_process_on_panic::tests::manual_thread --exact --nocapture
--ignored` with and without the new panic hook. The output is the same
except that the new hook prints some extra frames in the stack trace
that the default hook omits.