Driver `/solve` functionality part 4 (#958)
Progress on #895.

Implements mapping at the `boundary`.

Currently, to facilitate progress, the domain objects are not what I
want them to be ultimately. The domain model is somewhat
[anemic](https://en.wikipedia.org/wiki/Anemic_domain_model), but this
will be fixed as some of the interfaces become better-defined and as I
move the scoring logic into the `driver`. (It will be a while before
this; I will do this at some point as a refactoring _after_ I already
have automated tests for the `driver`).

The next step is calling `SettlementRater` from the `score` method.