Enable s3 instance uploading (#1027)
We want Quasimodo instances to be stored on AWS S3 for later debugging.
See #334.

This currently happens through the obsolete MIP solver which in our
cluster is configured to do nothing except upload the instance. However,
this instance uploading is hard to maintain and brittle. Recently it
started failing and filling up the disk space of the container. It is
also unnecessary overhead to run the whole MIP image just to upload
instances.

There were two main ideas for re-implementing the uploading in Rust:
1. Continue using a full solver pod to upload instances.
2. Integrate the uploading into the driver.

I went with 2. because the extra pod has too much organizational
overhead. I do not want to maintain a whole binary, pod, staging
configuration to upload instances. An even better solution might be to
do the instance uploading in Quasimodo because it isn't really relevant
to the driver. But that would be more work because it would be harder to
use Rust there.

To interface with S3 I decided on the official AWS Rust libraries. They
are technically in preview stage but this seems to be referring to
breaking changes and not bugs. I intentionally use as little of the API.
Breaking changes won't be difficult to handle.

I tried out manually interfacing with S3's http rest api in order to
reduce dependencies. This did not work well because AWS recently added
signing requirements to requests which are too difficult to implement
standalone. The AWS libraries we depend on now do not add many
transitive dependencies anyway because a lot of the tokio and http
libraries we already depend on.

Follow up tasks if this works on staging:
- Remove references to MIP solver in the Rust code.
- Remove references to MIP solver in the cluster.
- Clean up the old S3 bucket?