Refactor http solver instance creation (#1034)
After merging #1027 it turned out that the approach of uploading only
Quasimodo instances is not going to work because
- Quasimodo doesn't run in all environments.
- Quasimodo instances contain less information than other http solver
instances because non fee connected tokens are filtered out.

Instead, we want to always upload the "proper" instance. For every
auction one such instance is created and it is shared between all
solvers. This was handled by an ad hoc `InstanceCache` in
`http_solver.rs`. This implementation was hard to work with. In this PR
it becomes a less tightly coupled component in its own file
`instance_cache.rs`.

In order to do this I first needed to refactor the instance creation
code out of `http_solver.rs` because it should not be tied to single
solver. It now lives in `instance_creation.rs`.