Driver `/solve` tests (#1043)
Fixes https://github.com/cowprotocol/services/issues/895.

Successfully calling `/solve` on geth with access list simulation. (I.e.
I managed to figure out all the issues I was complaining about on the
weekly sync.) :tada:

What's done:

- A test for the `/solve` endpoint.
- A test that our local `example.solvers.toml` file is correct.
- Local testing environment based on dev-mode `geth`, I used `geth`
because it implements `eth_createAccessList` so I don't have to mock
access list simulation.
- A bunch of bugs fixed.
- Minor domain logic refactoring.
- A way to mock the current time in tests.
- A way to override contract addresses via CLI args.
- A way to disable access list simulation via CLI args.

What's missing:

- A way to reset the local geth env. Simply restarting the geth instance
can facilitate this, but there needs to be a way to do this
automatically at the start of each test.
- More test cases.
- Testing allowances?

I'm out of the shit for now, `/quote` and `/execute` should be fairly
easy to implement. Then when I get to implementing liquidity collection
I'll be in the shit again since it's a total unknown to me, it requires
a complex boundary integration, and I don't know how long it will take
or how to test it well.

Shoutout @MartinquaXD for serving as a sanity check at times when I was
stuck and @nlordell for giving me the pointer to follow the
`uniswapTrade` test from the contracts repo.

### Test Plan

Yyyyeeeeeeaaahhhhhh.