Get rid of deprecated solvable_orders routes (#435)
In the metrics I see that we still get requests to this route but in
Kibana I don't see any nginx logs. I think that means that some internal
service is requesting the route but I couldn't find it in this repo or
the kubernetes repo. If we continue to be unable to find it we can still
merge and hopefully see it logs afterwards.