Introduce Baseline Liquidity Domain Objects (#1045)
The first of a PR stack to port the Baseline solver to the new HTTP API
v2 solver model for solver + driver colocation.

This PR introduces the required liquidity domain types that the Baseline
solver will use. Specifically, constant product pools and weighted
pools.

In the following PRs, I will:
- [x] #1046
- [x] #1047
- [x] #1048
- [x] #1061
- [x] #1062
- [x] #1063

### Test Plan

Rust compiler. There will be an end-to-end test for the baseline solver
(an example was included in the test plan of #1039).