Do not group s3 uploads by date (#1065)
I've been thinking about allowing external access to the s3 instances
and came to the conclusion that it doesn't make sense to group them by
date. The Gnosis s3 script does something like this, probably so that
they can display the bucket like a folder structure without having too
many files in the same folder. However, to S3 there are no folders.
These are just files that happen to have slashes in their name. By
introducing some arbitrary partitioning scheme like the date, we make it
harder to know the path of an instance. By using only the id the path is
perfectly predictable.