Implement Baseline Solver Logic (#1061)
This PR adds the concrete Baseline solver implementation into the
`solvers` crate. Note that the actual path-finding logic is the existing
logic in the `shared` crate.

The implementation works by constructing a `boundary::baseline::Solver`
adapter that translates the new liquidity domain model into something
that can be used by the existing solver baseline logic. In the future, I
want to refactor `shared::baseline_solver` into the `solvers` crate, and
get rid of the `BaselineSolvable` trait - as it we won't need generics
anymore. Before that, I want a complete test-suite to make sure that we
don't introduce any regressions in the refactor.

### Test Plan

Rust compiler, E2E test incoming!