Gzip solver instance before S3 upload (#1086)
Follow up to
https://github.com/cowprotocol/services/pull/1027#discussion_r1062177319
.

Gzipping the instances makes them 90% of their original size which is a
large win for low effort. They can easily be gunzipped after
downloading. There are other choices for even better compression but I
didn't want to spend too much time on the choice and we can change it
later if needed.

I chose the flate2 crate because it is already in our dependency tree
through reqwest's gzip support.