Make auction update interval smarter (#1101)
Currently we wait a fixed interval of 2 seconds after every auction
update loop. This commit change the behavior to wait as long is needed
to reach the configured target effective update interval.

If creating the auction is slow then we sleep less or not at all. If
creating the auction is fast then we sleep the for longer.

This an improvement because we get more auctions on slow environments
and less on fast environments. For example, at the moment creating the
auction on staging Goerli is very fast which means there is a new
auction every 2 seconds. This creates unnecessary churn for the solvers
and for the data that we keep track of per auction like instance
uploading.

This PR makes the target interval configurable and sets the default to
10 seconds from the previous 2 seconds. This won't make auction creation
slower on prod mainnet because creating the auction already takes this
long.