Make driver run loop update interval smarter (#1104)
Same change as https://github.com/cowprotocol/services/pull/1101 but for
the driver. I forgot we'd need it there too.

As before, the reasoning for changing the default is that this won't
have any effect on busy networks where solving already takes 10 seconds.
Also need to adjust the value we set this too in the cluster, which I
have a Gitlab branch for.