Parallelize the driver tests (#1121)
Fixes https://github.com/cowprotocol/services/issues/1091.

Adds the ability for `dev-geth` to run multiple instances of `geth` on
different ports at the same time using docker host networking. Fixes two
nasty race conditions which lead to flaky tests:

- Not waiting for a block to be mined after publishing each transaction,
which could lead to two transactions getting mined in the same block,
which changes the gas price in the block where the settlement is
published and thereby changes the objective value causing a test to
fail.
- There was also a problem where the gas price used by the settlement in
the tests was not deterministic because of the way that the Uniswap V2
pool uses `block.timestamp` in the contract. This was fixed by waiting
for at least 1 second at a specific point during the test setup.

Fixes the `Dockerfile.dev-geth` which incorrectly used `COPY` instead of
`cp` in some places, this was missed before due to the way that
`docker-compose` caching works.

Adds a base datadir for the geth so that the seeded admin account and
genesis block are always the same.

### Test Plan

Automated tests.